[PDI-14690] Split web components into separate feature
