+++ pom.xml
@@ -48,6 +48,7 @@
   <modules>
     <module>api</module>
     <module>core</module>
+    <module>web</module>    
     <module>sample</module>
   </modules>
   <scm>
