+++ web/src/main/java/org/pentaho/metaverse/service/AnalyzerInfo.java
