+++ web/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -1,0 +1,48 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<blueprint xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
+           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+           xmlns:jaxrs="http://cxf.apache.org/blueprint/jaxrs"
+           xmlns:cxf="http://cxf.apache.org/blueprint/core"
+           xsi:schemaLocation="
+             http://www.osgi.org/xmlns/blueprint/v1.0.0 http://www.osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd
+             http://cxf.apache.org/blueprint/jaxrs http://cxf.apache.org/schemas/blueprint/jaxrs.xsd
+             http://cxf.apache.org/blueprint/core http://cxf.apache.org/schemas/blueprint/core.xsd
+             ">
+
+    <cxf:bus>
+        <cxf:features>
+            <cxf:logging/>
+        </cxf:features>
+    </cxf:bus>
+
+    <reference id="IMetaverseReader" interface="org.pentaho.metaverse.api.IMetaverseReader"/>
+    <reference id="IDocumentLocatorProvider" interface="org.pentaho.metaverse.api.IDocumentLocatorProvider"/>
+    <reference id="lineageCollector" interface="org.pentaho.metaverse.api.ILineageCollector"/>
+
+    <bean id="metaverseServiceBean" class="org.pentaho.metaverse.service.MetaverseService">
+        <argument ref="IMetaverseReader"/>
+        <argument ref="IDocumentLocatorProvider"/>
+        <property name="lineageCollector" ref="lineageCollector"/>
+    </bean>
+
+    <reference id="IStepAnalyzerProvider"
+               interface="org.pentaho.metaverse.api.analyzer.kettle.step.IStepAnalyzerProvider"/>
+    <reference id="IJobEntryAnalyzerProvider"
+               interface="org.pentaho.metaverse.api.analyzer.kettle.jobentry.IJobEntryAnalyzerProvider"/>
+
+    <bean id="analyzerInfoServiceBean" class="org.pentaho.metaverse.service.AnalyzerInfoService">
+        <property name="stepAnalyzerProvider" ref="IStepAnalyzerProvider"/>
+        <property name="jobEntryAnalyzerProvider" ref="IJobEntryAnalyzerProvider"/>
+    </bean>
+
+    <jaxrs:server id="metaverseService" address="/lineage">
+        <jaxrs:providers>
+            <bean class="org.codehaus.jackson.jaxrs.JacksonJaxbJsonProvider"/>
+        </jaxrs:providers>
+        <jaxrs:serviceBeans>
+            <ref component-id="metaverseServiceBean"/>
+            <ref component-id="analyzerInfoServiceBean"/>
+        </jaxrs:serviceBeans>
+    </jaxrs:server>
+
+</blueprint>
