+++ web/src/test/java/org/pentaho/metaverse/service/AnalyzerInfoServiceTest.java
