+++ web/src/test/java/org/pentaho/metaverse/service/AnalyzerInfoTest.java
